+++ csrc/aio/py_lib/deepspeed_py_io_handle.h
@@ -61,7 +61,7 @@ struct deepspeed_io_handle_t {
     int async_pwrite(const torch::Tensor& buffer, const char* filename);
 
     // TODO: Make API's args to be shape and dtype.
-    torch::Tensor new_cpu_locked_tensor(const size_t num_elem, const torch::Tensor& example_tensor);
+    torch::Tensor new_cpu_locked_tensor(const long long int num_elem, const torch::Tensor& example_tensor);
 
     bool free_cpu_locked_tensor(torch::Tensor&);
